+++ streamalert_cli/test/event.py
@@ -320,78 +320,3 @@ class TestEvent:
         # and update the test event with a full 'data' key
         default_test_event.update(self.override_record)
         self._event['data'] = default_test_event
-
-
-class TestEventFile:
-    """TestEventFile handles caching results of test events within a test file"""
-
-    def __init__(self, full_path, config):
-        self._full_path = full_path
-        self._config = config
-        self._results = []
-        self._error = None
-
-    def __bool__(self):
-        return bool(self._results)
-
-    def __str__(self):
-        output = [format_underline('\nFile: {file_name}\n'.format(file_name=self._full_path))]
-
-        if self._error:
-            output.append(self._error)
-        else:
-            for result in self._results:
-                output.append(result)
-
-        return '\n'.join(str(item) for item in output)
-
-    @property
-    def all_passed(self):
-        return self.passed == len(self._results)
-
-    @property
-    def passed(self):
-        return sum(1 for result in self._results if result.passed)
-
-    @property
-    def failed(self):
-        return sum(1 for result in self._results if not result.passed)
-
-    def add_result(self, result):
-        self._results.append(result)
-
-    def load_file(self):
-        """Helper to json load the contents of a file with some error handling
-
-        Test files should be formatted as:
-
-        [
-            {
-                "data": {},
-                "description": "...",
-                "...": "..."
-            }
-        ]
-
-        Returns:
-            dict: Loaded JSON from test event file
-        """
-        with open(self._full_path, 'r') as test_event_file:
-            try:
-                data = json.load(test_event_file)
-            except (ValueError, TypeError):
-                self._error = 'Test event file is not valid JSON'
-                return
-
-            if not isinstance(data, list):
-                self._error = 'Test event file is improperly formatted'
-                return
-
-            for idx, event in enumerate(data):
-                yield TestEvent(idx, event)
-
-    def process_file(self):
-        for test_event in self.load_file():
-            self._results.append(test_event)
-            if test_event.prepare(self._config):
-                yield test_event
